+++ CHANGELOG.md
@@ -8,6 +8,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 ### Added
+- All delegations list by address
+
+### Added
 - Added proxy api for delegates
 
 ## [0.1.9] - 2024-08-13